Isolate allocation-sensitive tests
Extract decompression into ParquetDecompressor; harden all remaining
reader crash sites found by AFL fuzzer:

- ParquetDecompressor: dedicated type for all codec decompression;
  uint expectedLength throughout; byte count check centralised; ZstdException
  caught and wrapped as CorruptParquetException
- PageHeader.UncompressedPageSize: uint (validated at parse time)
- DeltaBinaryPackedDecoder: add ReadUInt32 for length arrays (prefix/suffix
  lengths in delta-byte-array and delta-length-byte-array are always >= 0)
- ColumnChunkReader: guard level bytes vs compressed payload length;
  guard ReadBitPackedValue byteIndex vs payload; use uint for
  DecodeDictionaryIndexesIntoBuffer loop variables; fix ReadUnsignedVarInt
  empty-span crash; guard DecodeDeltaLengthByteArray per-entry length
- MemoryReadSource.ReadExactly: fix ulong underflow in bounds check
- ParquetReader.Reset: guard footer length vs stream size before subtraction

+++ Plank/Reading/PageHeaderReader.cs
@@ -9,7 +9,7 @@ static class PageHeaderReader
         var reader = new CompactProtocolReader(buffer);
         var previousFieldId = 0;
         var type = PageHeaderType.DataPage;
-        var uncompressedPageSize = 0;
+        var uncompressedPageSize = 0U;
         var compressedPageSize = 0;
         var valueCount = 0U;
         var encoding = EncodingKind.Plain;
@@ -26,7 +26,7 @@ static class PageHeaderReader
                     type = (PageHeaderType)reader.ReadI32();
                     break;
                 case 2:
-                    uncompressedPageSize = (int)reader.ReadI32AsU32(max: reader.Remaining);
+                    uncompressedPageSize = reader.ReadI32AsU32(max: reader.Remaining);
                     break;
                 case 3:
                     compressedPageSize = (int)reader.ReadI32AsU32(max: reader.Remaining);
